Update readme.md
This change to the readme is to instruct developers that use the `as2d` package that they cannot use the latest version of the AssemblyScript compiler with this software. This is because AssemblyScript experienced a very large backwards incompatible change that updated the memory model and introduced reference counting garbage collection.

In the future, there will be a backwards compatible rewrite that uses the `as-pect` testing framework. Every test will be rewritten, and the entire library will be written from scratch *again* to take advantage of all the new features AssemblyScript has to offer.  No breaking changes to the API will occur, and migration will be expected to be easy.